fixing build

+++ zhuravlev.ivan/I4/main.cpp
@@ -68,14 +68,14 @@ int main(int argc, char* argv[])
     {
       inputArraySize(n, m);
       int **arr = new int*[n];
-      for (int i = 0; i < n; ++i)
+      for (size_t i = 0; i < n; ++i)
       {
         arr[i] = new int[m];
       }
       fillDynArray(**arr, n, m, input);
       size_t size_dyn_arr;
       outputfile << maxSumDiag(**arr,size_dyn_arr);
-      for (int i = 0; i < n; ++i)
+      for (size_t i = 0; i < n; ++i)
       {
         delete[] arr[i];
       }
